chore : update model
1. ERD 토대로 변경되지 않은 부분 수정(reviews의 rating과
   orders의 quantity를 TINYINT로 수정, model의 관계 수정)

+++ migrations/20230717155347-create-order.js
@@ -45,7 +45,7 @@ module.exports = {
         defaultValue: false,
       },
       quantity: {
-        type: Sequelize.INTEGER,
+        type: Sequelize.TINYINT,
         allowNull: false,
         defaultValue: 1,
       },
